undid reactor read only and refactored
