+++ bluemira/codes/python_occ/_guard.py
@@ -1,0 +1,37 @@
+# SPDX-FileCopyrightText: 2021-present M. Coleman, J. Cook, F. Franza
+# SPDX-FileCopyrightText: 2021-present I.A. Maione, S. McIntosh
+# SPDX-FileCopyrightText: 2021-present J. Morris, D. Short
+#
+# SPDX-License-Identifier: LGPL-2.1-or-later
+
+"""
+Importer for external code API and related functions
+"""
+
+try:
+    import OCC  # noqa: F401
+
+    occ_available = True
+except ImportError:
+    occ_available = False
+
+
+def occ_available(f):
+    """
+    Check if the OCC module is available.
+
+    Raises
+    ------
+    ImportError
+        If the OCC module is not available.
+    """  # noqa: DOC201
+
+    def wrap(*args, **kwargs):
+        if not occ_available:
+            raise ImportError(
+                "OCC is not available. Run `conda install pythonocc-core` "
+                "to use this function."
+            )
+        return f(*args, **kwargs)
+
+    return wrap
