+++ eudemo/config/build_config.json
@@ -5,10 +5,10 @@
     "mixtures": "$path_expand:./materials_data/mixtures.json"
   },
   "Radial build": {
-    "run_mode": "read",
+    "run_mode": "run",
     "read_dir": "$path_expand:./",
     "run_dir": "$path_expand:./",
-    "plot": false
+    "plot": true
   },
   "Fixed boundary equilibrium": {
     "run_mode": "read",
@@ -51,7 +51,7 @@
     }
   },
   "Dummy fixed boundary equilibrium": {
-    "run_mode": "read",
+    "run_mode": "run",
     "param_class": "bluemira.equilibria.shapes::KuiroukidisLCFS",
     "shape_config": {
       "f_delta_l": 1.25,
@@ -64,7 +64,7 @@
     }
   },
   "Free boundary equilibrium": {
-    "run_mode": "read",
+    "run_mode": "run",
     "diagnostic_plotting": { "plot": ["eq", "convergence"] },
     "summary_plot": true,
     "save": true,
@@ -92,7 +92,7 @@
         "x1": { "value": "r_fw_ib_in" },
         "x2": { "value": "r_fw_ob_in" }
       },
-      "run_mode": "read",
+      "run_mode": "run",
       "file_path": "$path_expand:./FirstWallDesign.json",
       "problem_settings": { "n_koz_points": 100 },
       "optimisation_settings": {
@@ -130,7 +130,7 @@
     "show_data": false
   },
   "TF coils": {
-    "run_mode": "read",
+    "run_mode": "run",
     "file_path": "$path_expand:./TFCoilDesign.json",
     "plot": true,
     "material": {
@@ -156,7 +156,7 @@
     }
   },
   "PF coils": {
-    "run_mode": "read",
+    "run_mode": "run",
     "file_path": "$path_expand:./PFCoilDesign.json",
     "material": {
       "Ground Insulation": "Poloidal_Field_Coil",
