+++ eudemo/eudemo/reactor.py
@@ -581,9 +581,6 @@ if __name__ == "__main__":
         ivc_shapes.divertor_face,
     )
 
-    reactor.save_cad("xyz", {"without_components": [reactor.plasma]}, cad_format="dagmc")
-    exit()  # noqa: PLR1722
-
     upper_port_designer = UpperPortKOZDesigner(
         reactor_config.params_for("Upper Port"),
         reactor_config.config_for("Upper Port"),
